+++ src/Commands/CacheStoreCommand.php
@@ -141,7 +141,7 @@ class CacheStoreCommand extends CommandAbstract
      */
     protected function SkipIfCacheConfigIsEmpty(RunConfig $ciConfig)
     {
-        if (!$ciConfig->getCache()) {
+        if (!$ciConfig->getCache()->count()) {
             throw new SkipException('Cache config is empty. Skipping.');
         }
     }
@@ -183,7 +183,7 @@ class CacheStoreCommand extends CommandAbstract
             /** @var ConfigReader $configReader */
             $configReader = $this->container->get('trivago.rumi.services.config_reader');
 
-            return $configReader->getConfig($this->getWorkingDir(), $configFile);
+            return $configReader->getRunConfig($this->getWorkingDir(), $configFile);
         } catch (\Exception $e) {
             throw new \Exception('Required file \'' . $configFile . '\' does not exist');
         }