+++ src/Commands/Run/StageExecutor.php
@@ -28,7 +28,9 @@ use Trivago\Rumi\Events\JobFinishedEvent;
 use Trivago\Rumi\Events\JobStartedEvent;
 use Trivago\Rumi\Exceptions\CommandFailedException;
 use Trivago\Rumi\Models\JobConfig;
+use Trivago\Rumi\Models\JobConfigCollection;
 use Trivago\Rumi\Models\RunningCommand;
+use Trivago\Rumi\Models\StageConfig;
 use Trivago\Rumi\Models\VCSInfo\VCSInfoInterface;
 use Trivago\Rumi\Process\RunningProcessesFactory;
 
@@ -65,28 +67,30 @@ class StageExecutor
     }
 
     /**
-     * @param JobConfig[] $jobs
+     * @param StageConfig $stageConfig
      * @param $volume
-     * @param OutputInterface  $output
+     * @param OutputInterface $output
      * @param VCSInfoInterface $VCSInfo
      */
-    public function executeStage($jobs, $volume, OutputInterface $output, VCSInfoInterface $VCSInfo)
+    public function executeStage(StageConfig $stageConfig, $volume, OutputInterface $output, VCSInfoInterface $VCSInfo)
     {
-        $this->handleProcesses($output, $this->startStageProcesses($jobs, $VCSInfo, $volume));
+        $processes = $this->startStageProcesses($stageConfig->getJobs(), $VCSInfo, $volume);
+        $this->handleProcesses($output, $processes);
     }
 
     /**
-     * @param JobConfig[]      $jobs
-     * @param VCSInfoInterface $VCSInfo
+     * @param JobConfigCollection $jobConfigCollection
+     * @param VCSInfoInterface    $VCSInfo
      * @param $volume
      *
      * @return array
      */
-    private function startStageProcesses($jobs, VCSInfoInterface $VCSInfo, $volume)
+    private function startStageProcesses(JobConfigCollection $jobConfigCollection, VCSInfoInterface $VCSInfo, $volume)
     {
         $processes = [];
 
-        foreach ($jobs as $jobConfig) {
+        /** @var JobConfig $jobConfig */
+        foreach ($jobConfigCollection as $jobConfig) {
             $runningCommand = new RunningCommand(
                 $jobConfig,
                 $this->dockerComposeYamlBuilder->build($jobConfig, $VCSInfo, $volume),
@@ -131,7 +135,7 @@ class StageExecutor
                     $output->writeln(sprintf('<info>Executing job: %s</info>', $runningCommand->getJobName()));
                     $output->write($runningCommand->getOutput());
                     if (!empty($timeout)) {
-                        $output->writeln(PHP_EOL.'Process timed out after ' . $runningCommand->getTimeout().'s');
+                        $output->writeln(PHP_EOL.'Process timed out after '.$runningCommand->getTimeout().'s');
                     }
 
                     $this->dispatchJobFinishedEvent(