+++ src/Commands/RunCommand.php
@@ -21,15 +21,17 @@ namespace Trivago\Rumi\Commands;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
-use Trivago\Rumi\Builders\JobConfigBuilder;
+use Trivago\Rumi\Commands\Run\StageExecutor;
 use Trivago\Rumi\Events;
 use Trivago\Rumi\Events\RunFinishedEvent;
 use Trivago\Rumi\Events\RunStartedEvent;
 use Trivago\Rumi\Events\StageFinishedEvent;
 use Trivago\Rumi\Events\StageStartedEvent;
+use Trivago\Rumi\Models\RunConfig;
+use Trivago\Rumi\Models\StageConfig;
 use Trivago\Rumi\Models\VCSInfo\GitInfo;
+use Trivago\Rumi\Models\VCSInfo\VCSInfoInterface;
 use Trivago\Rumi\Services\ConfigReader;
 use Trivago\Rumi\Timer;
 
@@ -42,19 +44,9 @@ class RunCommand extends CommandAbstract
     const VOLUME = 'volume';
 
     /**
-     * @var string|null
-     */
-    private $volume;
-
-    /**
      * @var string
      */
     private $workingDir;
-
-    /**
-     * @var ContainerInterface
-     */
-    private $container;
 
     /**
      * @var EventDispatcherInterface
@@ -67,15 +59,24 @@ class RunCommand extends CommandAbstract
     private $configReader;
 
     /**
-     * @param ContainerInterface       $container
-     * @param EventDispatcherInterface $eventDispatcher
+     * @var StageExecutor
      */
-    public function __construct(ContainerInterface $container, EventDispatcherInterface $eventDispatcher)
-    {
+    private $stageExecutor;
+
+    /**
+     * @param EventDispatcherInterface $eventDispatcher
+     * @param ConfigReader             $configReader
+     * @param StageExecutor            $stageExecutor
+     */
+    public function __construct(
+        EventDispatcherInterface $eventDispatcher,
+        ConfigReader $configReader,
+        StageExecutor $stageExecutor
+) {
         parent::__construct();
-        $this->container = $container;
         $this->eventDispatcher = $eventDispatcher;
-        $this->configReader = $container->get('trivago.rumi.services.config_reader');
+        $this->configReader = $configReader;
+        $this->stageExecutor = $stageExecutor;
     }
 
     protected function configure()
@@ -116,64 +117,28 @@ class RunCommand extends CommandAbstract
     {
         try {
             if (trim($input->getArgument('volume')) != '') {
-                $this->volume = $input->getArgument(self::VOLUME);
+                $volume = $input->getArgument(self::VOLUME);
             } else {
-                $this->volume = $this->getWorkingDir();
+                $volume = $this->getWorkingDir();
             }
+
             $VCSInfo = new GitInfo(
                 $input->getArgument(self::GIT_URL),
                 $input->getArgument(self::GIT_COMMIT),
                 $input->getArgument(self::GIT_BRANCH)
             );
 
-            $timeTaken = Timer::execute(function () use ($input, $output, $VCSInfo) {
-                $runConfig = $this->configReader->getConfig($this->getWorkingDir(), $input->getOption(self::CONFIG));
+            $configFilePath = $input->getOption(self::CONFIG);
 
-                /** @var JobConfigBuilder $jobConfigBuilder */
-                $jobConfigBuilder = $this->container->get('trivago.rumi.job_config_builder');
+            $runConfig = $this->configReader->getRunConfig($this->getWorkingDir(), $configFilePath);
 
-                $this->eventDispatcher->dispatch(
-                    Events::RUN_STARTED, new RunStartedEvent($runConfig)
-                );
+            $this->eventDispatcher->dispatch(Events::RUN_STARTED, new RunStartedEvent($runConfig));
 
-                foreach ($runConfig->getStages() as $stageName => $stageConfig) {
-                    try {
-                        $jobs = $jobConfigBuilder->build($stageConfig);
-
-                        $this->eventDispatcher->dispatch(
-                            Events::STAGE_STARTED,
-                            new StageStartedEvent($stageName, $jobs)
-                        );
-
-                        $output->writeln(sprintf('<info>Stage: "%s"</info>', $stageName));
-
-                        $time = Timer::execute(
-                            function () use ($jobs, $output, $VCSInfo) {
-                                $this
-                                    ->container
-                                    ->get('trivago.rumi.commands.run.stage_executor')
-                                    ->executeStage($jobs, $this->volume, $output, $VCSInfo);
-                            }
-                        );
-
-                        $output->writeln('<info>Stage completed: '.$time.'</info>'.PHP_EOL);
-
-                        $this->eventDispatcher->dispatch(
-                            Events::STAGE_FINISHED,
-                            new StageFinishedEvent(StageFinishedEvent::STATUS_SUCCESS, $stageName)
-                        );
-                    } catch (\Exception $e) {
-                        $this->eventDispatcher->dispatch(
-                            Events::STAGE_FINISHED,
-                            new StageFinishedEvent(StageFinishedEvent::STATUS_FAILED, $stageName)
-                        );
-
-                        throw $e;
-                    }
-                }
-
-                $this->eventDispatcher->dispatch(Events::RUN_FINISHED, new RunFinishedEvent(RunFinishedEvent::STATUS_SUCCESS));
+            $timeTaken = Timer::execute(function () use ($runConfig, $output, $VCSInfo, $volume) {
+                $this->startRun($runConfig, $output, $VCSInfo, $volume);
             });
+
+            $this->eventDispatcher->dispatch(Events::RUN_FINISHED, new RunFinishedEvent(RunFinishedEvent::STATUS_SUCCESS));
 
             $output->writeln('<info>Build successful: '.$timeTaken.'</info>');
         } catch (\Exception $e) {
@@ -185,5 +150,49 @@ class RunCommand extends CommandAbstract
         }
 
         return 0;
+    }
+
+    /**
+     * @param RunConfig $runConfig
+     * @param OutputInterface $output
+     * @param VCSInfoInterface $VCSInfo
+     * @param string $volume
+     * @throws \Exception
+     *
+     * My intention is to move it to RunExecutor class
+     */
+    private function startRun(RunConfig $runConfig, OutputInterface $output, VCSInfoInterface $VCSInfo, string $volume){
+        foreach ($runConfig->getStagesCollection() as $stage) {
+            try {
+                $this->eventDispatcher->dispatch(
+                    Events::STAGE_STARTED,
+                    new StageStartedEvent($stage)
+                );
+
+                $output->writeln(sprintf('<info>Stage: "%s"</info>', $stage->getName()));
+
+                $time = Timer::execute(
+                    function () use ($stage, $output, $VCSInfo, $volume) {
+                        $this->stageExecutor->executeStage($stage, $volume, $output, $VCSInfo);
+                    }
+                );
+
+                $output->writeln('<info>Stage completed: '.$time.'</info>'.PHP_EOL);
+
+                $this->eventDispatcher->dispatch(
+                    Events::STAGE_FINISHED,
+                    new StageFinishedEvent(StageFinishedEvent::STATUS_SUCCESS, $stage)
+                );
+            } catch (\Exception $e) {
+                $this->eventDispatcher->dispatch(
+                    Events::STAGE_FINISHED,
+                    new StageFinishedEvent(StageFinishedEvent::STATUS_FAILED, $stage)
+                );
+
+                throw $e;
+            }
+        }
+
+
     }
 }