+++ src/Services/ConfigReader.php
@@ -19,12 +19,28 @@
 namespace Trivago\Rumi\Services;
 
 use Symfony\Component\Yaml\Parser;
+use Trivago\Rumi\Builders\JobConfigBuilder;
 use Trivago\Rumi\Commands\ReturnCodes;
+use Trivago\Rumi\Models\CacheConfig;
 use Trivago\Rumi\Models\RunConfig;
+use Trivago\Rumi\Models\StagesCollection;
 
 class ConfigReader
 {
     const CONFIG_FILE = '.rumi.yml';
+    /**
+     * @var JobConfigBuilder
+     */
+    private $jobConfigBuilder;
+
+    /**
+     * ConfigReader constructor.
+     * @param JobConfigBuilder $jobConfigBuilder
+     */
+    public function __construct(JobConfigBuilder $jobConfigBuilder)
+    {
+        $this->jobConfigBuilder = $jobConfigBuilder;
+    }
 
     /**
      * @param $workingDir
@@ -34,7 +50,7 @@ class ConfigReader
      *
      * @return RunConfig
      */
-    public function getConfig($workingDir, $configFile)
+    public function getRunConfig($workingDir, $configFile)
     {
         $configFilePath = $workingDir . $configFile;
 
@@ -49,9 +65,12 @@ class ConfigReader
         $ciConfig = $parser->parse(file_get_contents($configFilePath));
 
         return new RunConfig(
-            !empty($ciConfig['stages']) ? $ciConfig['stages'] : [],
-            !empty($ciConfig['cache']) ? $ciConfig['cache'] : [],
-            !empty($ciConfig['merge_branch']) ? $ciConfig['merge_branch'] : null
+            new StagesCollection(
+                $this->jobConfigBuilder,
+                $ciConfig['stages'] ?? []
+            ),
+            new CacheConfig($ciConfig['cache'] ?? []),
+            $ciConfig['merge_branch'] ?? ""
         );
     }
 }