+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
+- Fix stderr and stdout ordering in the docker-compose logs output
+- Migrated from php 5 to php 7
+- Updated all composer dependencies
 
 ### Removed